Add files via upload
so this is the 2.1 version of the file. it's not the first version, not by a long shot, but it does have all the features I've got so far:
Automatic module power editing, Manual module power editing, Automatic Item chance edit based on power tokens, Automatic Weapon chance edit based on power tokens,
automatic combined item and weapon chance edit to return chances to normal, a static OP module edit for my own preferences, and a way to open the save file directly in your default text editor.
There's little to no documentation for this file, and i hope to change that with future updates.